Recorded merge of revisions 28475 via svnmerge from svn+ssh://uwog@svn.abisource.com/svnroot/abiword/trunk
........
  r28475 | jbrefort | 2009-11-21 16:33:45 +0100 (Sat, 21 Nov 2009) | 2 lines
  
  Fixed after one more goffice API change, please backport.
........


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/branches/ABI-2-8-0-STABLE@28644 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ plugins/goffice/unix/AbiGOffice.cpp
@@ -76,7 +76,7 @@ abi_set_sensitive (G_GNUC_UNUSED GOCmdContext *cc,
 }
 
 static void
-abi_progress_set (G_GNUC_UNUSED GOCmdContext *cc, G_GNUC_UNUSED gfloat val)
+abi_progress_set (G_GNUC_UNUSED GOCmdContext *cc, G_GNUC_UNUSED double val)
 {
 }
